Закрытие от публичного доступа

+++ backend/controllers/ContentController.php
@@ -36,6 +36,10 @@ class ContentController extends Controller
      */
     public function actionIndex()
     {
+        if (!Yii::$app->user->id){
+            return $this->goHome();
+        }
+
         $searchModel = new SearchContent();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -53,6 +57,10 @@ class ContentController extends Controller
      */
     public function actionView($id)
     {
+        if (!Yii::$app->user->id){
+            return $this->goHome();
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -65,6 +73,9 @@ class ContentController extends Controller
      */
     public function actionCreate()
     {
+        if (!Yii::$app->user->id){
+            return $this->goHome();
+        }
         $model = new Content();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -85,6 +96,10 @@ class ContentController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (!Yii::$app->user->id){
+            return $this->goHome();
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -105,6 +120,10 @@ class ContentController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!Yii::$app->user->id){
+            return $this->goHome();
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
@@ -119,6 +138,9 @@ class ContentController extends Controller
      */
     protected function findModel($id)
     {
+        if (!Yii::$app->user->id){
+            return $this->goHome();
+        }
         if (($model = Content::findOne($id)) !== null) {
             return $model;
         }
